tests: Move RunVMIAndExpectLaunchIgnoreWarnings to libclient
