+++ tests/network/probes.go
@@ -6,6 +6,7 @@ import (
 	"time"
 
 	"kubevirt.io/kubevirt/tests/framework/kubevirt"
+	"kubevirt.io/kubevirt/tests/libclient"
 
 	expect "github.com/google/goexpect"
 	. "github.com/onsi/ginkgo/v2"
@@ -109,7 +110,7 @@ var _ = SIGDescribe("[ref_id:1182]Probes", func() {
 				By(specifyingVMReadinessProbe)
 				vmi = libvmifact.NewFedora(
 					withMasqueradeNetworkingAndFurtherUserConfig(withReadinessProbe(readinessProbe))...)
-				vmi = tests.RunVMIAndExpectLaunchIgnoreWarnings(vmi, 180)
+				vmi = libclient.RunVMIAndExpectLaunchIgnoreWarnings(vmi, 180)
 
 				By("Waiting for agent to connect")
 				Eventually(matcher.ThisVMI(vmi), 12*time.Minute, 2*time.Second).Should(matcher.HaveConditionTrue(v1.VirtualMachineInstanceAgentConnected))
@@ -134,7 +135,7 @@ var _ = SIGDescribe("[ref_id:1182]Probes", func() {
 					withMasqueradeNetworkingAndFurtherUserConfig(
 						withReadinessProbe(
 							createGuestAgentPingProbe(period, initialSeconds)))...)
-				vmi = tests.RunVMIAndExpectLaunchIgnoreWarnings(vmi, 180)
+				vmi = libclient.RunVMIAndExpectLaunchIgnoreWarnings(vmi, 180)
 				By("Waiting for agent to connect")
 				Eventually(matcher.ThisVMI(vmi), 12*time.Minute, 2*time.Second).Should(matcher.HaveConditionTrue(v1.VirtualMachineInstanceAgentConnected))
 
@@ -160,7 +161,7 @@ var _ = SIGDescribe("[ref_id:1182]Probes", func() {
 		DescribeTable("should fail", func(readinessProbe *v1.Probe, vmiFactory func(opts ...libvmi.Option) *v1.VirtualMachineInstance) {
 			By(specifyingVMReadinessProbe)
 			vmi = vmiFactory(withReadinessProbe(readinessProbe))
-			vmi = tests.RunVMIAndExpectLaunchIgnoreWarnings(vmi, 180)
+			vmi = libclient.RunVMIAndExpectLaunchIgnoreWarnings(vmi, 180)
 
 			By("Checking that the VMI is consistently non-ready")
 			Consistently(matcher.ThisVMI(vmi), 30*time.Second, 100*time.Millisecond).Should(matcher.HaveConditionMissingOrFalse(v1.VirtualMachineInstanceReady))
@@ -211,7 +212,7 @@ var _ = SIGDescribe("[ref_id:1182]Probes", func() {
 				vmi = libvmifact.NewFedora(
 					withMasqueradeNetworkingAndFurtherUserConfig(
 						withLivelinessProbe(livenessProbe))...)
-				vmi = tests.RunVMIAndExpectLaunchIgnoreWarnings(vmi, 180)
+				vmi = libclient.RunVMIAndExpectLaunchIgnoreWarnings(vmi, 180)
 
 				By("Waiting for agent to connect")
 				Eventually(matcher.ThisVMI(vmi), 12*time.Minute, 2*time.Second).Should(matcher.HaveConditionTrue(v1.VirtualMachineInstanceAgentConnected))
@@ -236,7 +237,7 @@ var _ = SIGDescribe("[ref_id:1182]Probes", func() {
 				vmi = libvmifact.NewFedora(withMasqueradeNetworkingAndFurtherUserConfig(
 					withLivelinessProbe(createGuestAgentPingProbe(period, initialSeconds)),
 				)...)
-				vmi = tests.RunVMIAndExpectLaunchIgnoreWarnings(vmi, 180)
+				vmi = libclient.RunVMIAndExpectLaunchIgnoreWarnings(vmi, 180)
 
 				By("Waiting for agent to connect")
 				Eventually(matcher.ThisVMI(vmi), 12*time.Minute, 2*time.Second).Should(matcher.HaveConditionTrue(v1.VirtualMachineInstanceAgentConnected))
@@ -255,7 +256,7 @@ var _ = SIGDescribe("[ref_id:1182]Probes", func() {
 		DescribeTable("should fail the VMI", func(livenessProbe *v1.Probe, vmiFactory func(opts ...libvmi.Option) *v1.VirtualMachineInstance) {
 			By("Specifying a VMI with a livenessProbe probe")
 			vmi = vmiFactory(withLivelinessProbe(livenessProbe))
-			vmi = tests.RunVMIAndExpectLaunchIgnoreWarnings(vmi, 180)
+			vmi = libclient.RunVMIAndExpectLaunchIgnoreWarnings(vmi, 180)
 
 			By("Checking that the VMI is in a final state after a while")
 			Eventually(func() bool {
@@ -298,14 +299,14 @@ func createReadyAlpineVMIWithReadinessProbe(probe *v1.Probe) *v1.VirtualMachineI
 	vmi := libvmifact.NewAlpineWithTestTooling(
 		withMasqueradeNetworkingAndFurtherUserConfig(withReadinessProbe(probe))...)
 
-	return tests.RunVMIAndExpectLaunchIgnoreWarnings(vmi, 180)
+	return libclient.RunVMIAndExpectLaunchIgnoreWarnings(vmi, 180)
 }
 
 func createReadyAlpineVMIWithLivenessProbe(probe *v1.Probe) *v1.VirtualMachineInstance {
 	vmi := libvmifact.NewAlpineWithTestTooling(
 		withMasqueradeNetworkingAndFurtherUserConfig(withLivelinessProbe(probe))...)
 
-	return tests.RunVMIAndExpectLaunchIgnoreWarnings(vmi, 180)
+	return libclient.RunVMIAndExpectLaunchIgnoreWarnings(vmi, 180)
 }
 
 func createTCPProbe(period int32, initialSeconds int32, port int) *v1.Probe {
